+++ internal/parser/parser_test.go
@@ -2754,6 +2754,22 @@ func TestLowerOptionalChain(t *testing.T) {
 	// Check that direct eval status is propagated through optional chaining
 	expectPrintedTarget(t, 2019, "eval?.(x)", "eval == null ? void 0 : eval(x);\n")
 	expectPrintedTarget(t, 2019, "(1 ? eval : 0)?.(x)", "eval == null ? void 0 : (0, eval)(x);\n")
+
+	// Check super property access
+	expectPrintedTarget(t, 2019, "class Foo extends Bar { foo() { super.bar?.() } }", `class Foo extends Bar {
+  foo() {
+    var _a;
+    (_a = super.bar) == null ? void 0 : _a.call(this);
+  }
+}
+`)
+	expectPrintedTarget(t, 2019, "class Foo extends Bar { foo() { super['bar']?.() } }", `class Foo extends Bar {
+  foo() {
+    var _a;
+    (_a = super["bar"]) == null ? void 0 : _a.call(this);
+  }
+}
+`)
 }
 
 func TestLowerOptionalCatchBinding(t *testing.T) {